laptop/requirements.txt
paho-mqtt
numpy
scikit-learn
joblib
flask
flask-cors
python-dotenv